Move BoundaryPath to common.rs
