+++ homotopy-core/src/common.rs
@@ -295,6 +295,45 @@ impl<T> IndexMut<SliceIndex> for Vec<T> {
     }
 }
 
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
+#[cfg_attr(feature = "fuzz", derive(arbitrary::Arbitrary))]
+pub struct BoundaryPath(pub Boundary, pub usize);
+
+impl BoundaryPath {
+    pub fn split(path: &[SliceIndex]) -> (Option<Self>, Vec<Height>) {
+        use SliceIndex::{Boundary, Interior};
+
+        let mut boundary_path: Option<Self> = None;
+        let mut interior = Vec::new();
+
+        for height in path.iter().rev() {
+            match (&mut boundary_path, height) {
+                (Some(bp), _) => bp.1 += 1,
+                (None, Boundary(b)) => boundary_path = Some(Self(*b, 0)),
+                (None, Interior(h)) => interior.insert(0, *h),
+            }
+        }
+
+        (boundary_path, interior)
+    }
+
+    #[inline]
+    pub fn boundary(self) -> Boundary {
+        self.0
+    }
+
+    #[inline]
+    pub fn depth(self) -> usize {
+        self.1
+    }
+}
+
+impl From<Boundary> for BoundaryPath {
+    fn from(boundary: Boundary) -> Self {
+        Self(boundary, 0)
+    }
+}
+
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 #[cfg_attr(feature = "fuzz", derive(arbitrary::Arbitrary))]
 pub enum Direction {
